requirements.txt
paddleocr==2.6.1.3
paddlepaddle==2.5.2
sqlalchemy
psycopg2-binary
python-dotenv
tqdm
opencv-python
PyMuPDF
numpy==1.23.5
pytest